Conflicts:
	src/osgEarth/StringUtils
	src/osgEarthDrivers/engine_osgterrain/SinglePassTerrainTechnique.cpp
	src/osgEarthFeatures/SubstituteModelFilter.cpp
	src/osgEarthSymbology/Style.cpp

+++ src/osgEarth/CompositeTileSource.cpp
@@ -203,8 +203,8 @@ CompositeTileSource::createImage(const TileKey&        key,
             }
 
             // check that this source is within the level bounds:
-            if (minLevel > key.getLevelOfDetail() ||
-                maxLevel < key.getLevelOfDetail() )
+            if (minLevel > (int)key.getLevelOfDetail() ||
+                maxLevel < (int)key.getLevelOfDetail() )
             {
                 continue;
             }
